Server Testing
Its a refactoring errorHandler kind of day!
